refactor(quick-filters): extract checkbox query algebra into pure module
